AI-Marketing/requirements.txt
Python==3.11.4
streamlit==1.25.0
openai==0.27.8
Pillow==8.3.1 #just incase 

